+++ lib/json_web_token/jwt.rb
@@ -4,6 +4,7 @@ module JsonWebToken
   # Encode claims for transmission as a JSON object that is used as the payload of a JSON Web
   # Signature (JWS) structure, enabling the claims to be integrity protected with a Message
   # Authentication Code (MAC), to be later verified
+  # @see http://tools.ietf.org/html/rfc7519
   module Jwt
 
     ALG_DEFAULT = 'HS256'
@@ -19,12 +20,12 @@ module JsonWebToken
     #   (e.g String for Hmac | OpenSSL::PKey::RSA | OpenSSL::PKey::EC)
     # @return [String] a JSON Web Token, representing digitally signed claims
     # @example
-    #   claims = {iss: 'joe', exp: 1300819380, 'http://example.com/is_root' => true}
+    #   claims = {iss: 'joe', exp: 1300819380, :'http://example.com/is_root' => true}
     #   options = {alg: 'HS256', key: 'gZH75aKtMN3Yj0iPS4hcgUuTwjAzZr9C'}
     #   Jwt.sign(claims, options)
     #   # => 'eyJhbGciOiJIUzI1NiJ9.cGF5bG9hZA.uVTaOdyzp_f4mT_hfzU8LnCzdmlVC4t2itHDEYUZym4'
     # @see http://tools.ietf.org/html/rfc7519#section-7.1
-    def sign(claims, options = {})
+    def sign(claims, options)
       message = validated_message(claims)
       header = config_header(options)
       return Jws.unsecured_message(header, message) if header[:alg] == 'none'
@@ -33,14 +34,14 @@ module JsonWebToken
 
     # @param jwt [String] a JSON Web Token
     # @param options [Hash] specify the desired verifying algorithm and verifying key
-    # @return [Hash] a JWT claims set if the jwt verifies, or +{error: 'Invalid'}+ otherwise
+    # @return [Hash] a JWT claims set if the jwt verifies, or +error: 'Invalid'+ otherwise
     # @example
     #   jwt = 'eyJhbGciOiJIUzI1NiJ9.cGF5bG9hZA.uVTaOdyzp_f4mT_hfzU8LnCzdmlVC4t2itHDEYUZym4'
     #   options = {alg: 'HS256', key: 'gZH75aKtMN3Yj0iPS4hcgUuTwjAzZr9C'}
     #   Jwt.verify(jwt, options)
-    #   # => {iss: 'joe', exp: 1300819380, 'http://example.com/is_root' => true}
+    #   # => {iss: 'joe', exp: 1300819380, :'http://example.com/is_root' => true}
     # @see see http://tools.ietf.org/html/rfc7519#section-7.2
-    def verify(jwt, options = {})
+    def verify(jwt, options)
       alg = options[:alg] || ALG_DEFAULT
       jws = Jws.verify(jwt, alg, options[:key])
       jws ? Util.symbolize_keys(decoded_message_json_to_hash jws) : {error: 'invalid'}